menu/requirements.txt
asgiref==3.4.1
astroid==2.3.3

autopep8==1.5.7
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
black==21.8b0
bleach==4.1.0
blinker==1.4
bootstrap4==0.1.0
build==0.6.0.post1
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
commonmark==0.9.1
defusedxml==0.7.1
distlib==0.3.3
distro==1.4.0
dj-database-url==0.5.0
Django==3.2.7
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
django-filter==2.4.0
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-social-auth==0.7.28
django-social-auth3==0.7.23
djangorestframework==3.12.4
docutils==0.17.1
easy-thumbnails==2.7.1
entrypoints==0.3
fasteners==0.14.1
filelock==3.2.1
future==0.18.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.8.1
isort==4.3.4
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
logilab-common==1.4.3
loguru==0.5.3
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
MouseInfo==0.1.3
my-simple-split==0.0.1
myownsample==0.0.1
mypy-extensions==0.4.3
mysimplesplit==0.0.1
netifaces==0.10.4
oauth2==1.9.0.post1
oauth2.3==1.5.211
oauthlib==3.1.0
olefile==0.46
packaging==21.0
paramiko==2.6.0
pathspec==0.9.0
pbr==5.6.0
pep517==0.11.0
pexpect==4.6.0
Pillow==7.0.0
pkginfo==1.7.1
platformdirs==2.3.0
ProSplit==0.0.1
PyAutoGUI==0.9.53
pycairo==1.16.2
pycodestyle==2.7.0
pycrypto==2.6.1
Pygments==2.10.0
pyinotify==0.9.6
PyJWT==2.1.0
pylint==2.4.4
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.14.1
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.7.3
python-Levenshtein==0.12.0
python-openid==2.2.5
python-openid3==2.2.6
python3-openid==3.2.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2021.3
pywhatkit==5.1
pyxdg==0.26
PyYAML==5.3.1
readme-renderer==29.0
regex==2021.8.28
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==10.9.0
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.16.0
six==1.14.0
social==0.1.dev3
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
speedtest-cli==2.1.3
SplitPro==0.0.1
sqlparse==0.4.2
systemd-python==234
testresources==2.0.1
toml==0.10.2
tomli==1.2.1
tqdm==4.62.2
twine==3.4.2
typing-extensions==3.10.0.2
ultrasplit==0.0.1
urllib3==1.25.8
virtualenv==20.8.1
wadllib==1.3.3
webencodings==0.5.1
websocket-client==0.53.0
whitenoise==5.3.0
wikipedia==1.4.0
wrapt==1.11.2
zipp==3.5.0
